Improve input parameters and response quantity limits
Default return form the email check was all messages, which in some
cases will simply freeze the return. This ensures that only 50 messages
are evaluated and notifies the user of this message count.
Some grammar errors were corrected during the update of possible
commands and the addition of variable responses for commands that don't
require a static answer.