Begin header restoration process after determining places
Placement of the code was required, along with verification, before
finalizing where the plugins would be located. Now that these aspects
have been finalized, it is time to restore the original headers to
replace the temporary "CYA" notes.
All plugins that contain just the name of the plugin and no additional
information are not yet updated, but the others now include proper
header information from original source.