Design updated Wordnik plugin, update all outputs
The original Wordnik plugin was designed for deprecated editions of the
Wordnik library and resulted in an error stating the module could not
be found. This led to the discovery that the full plugin required an
updated design.
Additional outputs have been cleaned up to avoid redundant output of
words when they are being defined.